fixed the bug of Edit.jsx. There was a small mistake. I didn't pass arguments that the function onSubmit was expecting in Create.jsx and Edit.jsx

+++ app/Http/Controllers/ProjectController.php
@@ -115,29 +115,21 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project)
     {
-        dd($request);
-        // Validate and retrieve the validated data from the request
         $data = $request->validated();
-
-        // Get the image from the validated data or set it to null
         $image = $data['image'] ?? null;
-        
-        // Set the ID of the user making the update
         $data['updated_by'] = Auth::id();
 
-        // If an image is provided, store it and set the path
         if ($image) {
-            // Store the image in a unique subdirectory and get the path
+            if($project->image_path){
+                if (Storage::disk('public')->exists($project->image_path)) {
+                    Storage::disk('public')->delete($project->image_path);
+                }
+            }
             $data['image_path'] = $image->store('projects/' . Str::random(16), 'public');
         } else {
-            // If no new image is uploaded, retain the old image path
             $data['image_path'] = $project->image_path;
         }
-
-        // Update the project with the validated data
         $project->update($data);
-
-        // Redirect back with a success message
         return redirect()->route('projects.index')->with('success', 'Project updated successfully');
     }
 
@@ -147,8 +139,10 @@ class ProjectController extends Controller
     public function destroy(Project $project)
     {
         $name = $project->name;
-        if (Storage::disk('public')->exists($project->image_path)) {
-            Storage::disk('public')->delete($project->image_path);
+        if($project->image_path){
+            if (Storage::disk('public')->exists($project->image_path)) {
+                Storage::disk('public')->delete($project->image_path);
+            }
         }
         $project->delete();
         return redirect()->route('projects.index')->with('success', 'Project "'.$name.'" deleted successfully');
